requirements.txt
dnspython
Flask
Flask-SQLAlchemy
requests
sqlalchemy
python-whois
ip-whois
numpy